@@ -25,10 +25,11 @@ all = ["psycopg2"]
 dev = ["psycopg2-binary"]
 
 [tool.poetry.dependencies]
-python = ">=3.9,<3.12"
+python = ">=3.11,<3.12"
 django = ">=4.2,<4.3"
 djangorestframework = "*"
-dynaconf = ">=3"
+djangorestframework_simplejwt = "*"
+dynaconf = ">=3.1"
 drf-spectacular = "*"
 channels = { version = "*", extras = ["daphne"] }
 psycopg2 = { version = "*", optional = true }
@@ -41,11 +42,13 @@ cryptography = "*"
 kubernetes = "*"
 podman = ">=4.5"
 rq-scheduler = "^0.10"
-# Experimental LDAP Integration https://issues.redhat.com/browse/AAP-16938
-# We are temporarily updating it to use latest from devel branch to keep consistency accross
-# other AAP components
-django-ansible-base = { git = "https://github.com/ansible/django-ansible-base.git", branch="devel", extras=["authentication"] }
+django-ansible-base = { git = "https://github.com/ansible/django-ansible-base.git", tag = "2024.2.12", extras = [
+    "authentication",
+    "channel_auth",
+] }
 jinja2 = "*"
+django-split-settings = "^1.2.0"
+ansible-core = "^2.16.2"
 
 [tool.poetry.group.test.dependencies]
 pytest = "*"
