@@ -18,7 +18,7 @@ from abc import ABC, abstractmethod
 from datetime import datetime
 
 from django.conf import settings
-from pydantic import BaseModel
+from pydantic import BaseModel, validator
 
 import aap_eda.services.activation.engine.exceptions as exceptions
 from aap_eda.core.enums import ActivationStatus
@@ -50,20 +50,33 @@ class LogHandler(ABC):
 class AnsibleRulebookCmdLine(BaseModel):
     ws_url: str
     ws_ssl_verify: str
+    ws_access_token: str
+    ws_refresh_token: str
+    ws_token_url: str
     heartbeat: int
-    id: str
+    id: tp.Union[str, int]  # casted to str
     log_level: tp.Optional[str] = None  # -v or -vv or None
+
+    @validator("id", pre=True)
+    def cast_to_str(cls, v):
+        return str(v)
 
     def command(self) -> str:
         return "ansible-rulebook"
 
-    def get_args(self) -> list[str]:
+    def get_args(self, sanitized=False) -> list[str]:
         args = [
             "--worker",
             "--websocket-ssl-verify",
             self.ws_ssl_verify,
-            "--websocket-address",
+            "--websocket-url",
             self.ws_url,
+            "--websocket-access-token",
+            "******" if sanitized else self.ws_access_token,
+            "--websocket-refresh-token",
+            "******" if sanitized else self.ws_refresh_token,
+            "--websocket-token-url",
+            self.ws_token_url,
             "--id",
             self.id,
             "--heartbeat",
@@ -74,8 +87,8 @@ class AnsibleRulebookCmdLine(BaseModel):
 
         return args
 
-    def command_and_args(self) -> list[str]:
-        args = self.get_args()
+    def command_and_args(self, sanitized=False) -> list[str]:
+        args = self.get_args(sanitized)
         args.insert(0, self.command())
         return args
 
@@ -95,7 +108,7 @@ class ContainerRequest(BaseModel):
     ports: tp.Optional[list[tuple]] = None
     pull_policy: str = settings.DEFAULT_PULL_POLICY  # Always by default
     mem_limit: tp.Optional[str] = None
-    mounts: tp.Optional[dict] = None
+    mounts: tp.Optional[list[dict]] = None
     env_vars: tp.Optional[dict] = None
     extra_args: tp.Optional[dict] = None
 
