@@ -1,4 +1,4 @@
-#  Copyright 2023 Red Hat, Inc.
+#  Copyright 2024 Red Hat, Inc.
 #
 #  Licensed under the Apache License, Version 2.0 (the "License");
 #  you may not use this file except in compliance with the License.
@@ -12,4 +12,24 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-VALID_LOG_LEVELS = ("-v", "-vv")
+from ansible_base.resource_registry.registry import (
+    ResourceConfig,
+    ServiceAPIConfig,
+    SharedResource,
+)
+from ansible_base.resource_registry.shared_types import UserType
+
+from aap_eda.core import models
+
+
+class APIConfig(ServiceAPIConfig):
+    service_type = "eda"
+
+
+RESOURCE_LIST = (
+    ResourceConfig(
+        models.User,
+        shared_resource=SharedResource(serializer=UserType, is_provider=False),
+        name_field="username",
+    ),
+)
