@@ -11,7 +11,10 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-from ansible_base.urls import urls as base_urls
+from ansible_base.lib.dynamic_config.dynamic_urls import api_version_urls
+from ansible_base.resource_registry.urls import (
+    urlpatterns as resource_api_urls,
+)
 from django.urls import include, path
 from drf_spectacular.views import (
     SpectacularJSONAPIView,
@@ -20,6 +23,7 @@ from drf_spectacular.views import (
     SpectacularYAMLAPIView,
 )
 from rest_framework import routers
+from rest_framework_simplejwt import views as jwt_views
 
 from . import views
 
@@ -30,11 +34,11 @@ router.register("rulebooks", views.RulebookViewSet)
 router.register("rulesets", views.RulesetViewSet)
 router.register("rules", views.RuleViewSet)
 router.register("roles", views.RoleViewSet)
-router.register("tasks", views.TaskViewSet, basename="task")
 router.register("activations", views.ActivationViewSet)
 router.register("activation-instances", views.ActivationInstanceViewSet)
 router.register("audit-rules", views.AuditRuleViewSet)
 router.register("users", views.UserViewSet)
+router.register("event-streams", views.EventStreamViewSet)
 router.register(
     "users/me/awx-tokens",
     views.CurrentUserAwxTokenViewSet,
@@ -67,13 +71,20 @@ openapi_urls = [
 ]
 
 v1_urls = [
+    # <- Experimental LDAP Auth https://issues.redhat.com/browse/AAP-16938
+    path("", include(api_version_urls)),
+    # ->
+    path("", include(resource_api_urls)),
     path("", include(openapi_urls)),
     path("auth/session/login/", views.SessionLoginView.as_view()),
     path("auth/session/logout/", views.SessionLogoutView.as_view()),
+    path(
+        "auth/token/refresh/",
+        jwt_views.TokenRefreshView.as_view(),
+        name="token_refresh",
+    ),
     path("users/me/", views.CurrentUserView.as_view()),
     *router.urls,
-    # Experimental LDAP Integration https://issues.redhat.com/browse/AAP-16938
-    path("", include(base_urls)),
 ]
 
 urlpatterns = [
