@@ -28,12 +28,17 @@ from rest_framework.response import Response
 from aap_eda.api import exceptions as api_exc, filters, serializers
 from aap_eda.api.serializers.activation import is_activation_valid
 from aap_eda.core import models
-from aap_eda.core.enums import Action, ActivationStatus, ResourceType
+from aap_eda.core.enums import (
+    Action,
+    ActivationStatus,
+    ProcessParentType,
+    ResourceType,
+)
 from aap_eda.tasks.orchestrator import (
-    delete_activation,
-    restart_activation,
-    start_activation,
-    stop_activation,
+    delete_rulebook_process,
+    restart_rulebook_process,
+    start_rulebook_process,
+    stop_rulebook_process,
 )
 
 logger = logging.getLogger(__name__)
@@ -78,13 +83,16 @@ class ActivationViewSet(
         )
         serializer.is_valid(raise_exception=True)
 
-        response = serializer.create(serializer.validated_data)
+        activation = serializer.create(serializer.validated_data)
 
-        if response.is_enabled:
-            start_activation(activation_id=response.id)
+        if activation.is_enabled:
+            start_rulebook_process(
+                process_parent_type=ProcessParentType.ACTIVATION,
+                id=activation.id,
+            )
 
         return Response(
-            serializers.ActivationReadSerializer(response).data,
+            serializers.ActivationReadSerializer(activation).data,
             status=status.HTTP_201_CREATED,
         )
 
@@ -120,7 +128,10 @@ class ActivationViewSet(
         activation.status = ActivationStatus.DELETING
         activation.save(update_fields=["status"])
         logger.info(f"Now deleting {activation.name} ...")
-        delete_activation(activation_id=activation.id)
+        delete_rulebook_process(
+            process_parent_type=ProcessParentType.ACTIVATION,
+            id=activation.id,
+        )
 
     @extend_schema(
         description="List all instances for the Activation",
@@ -141,7 +152,7 @@ class ActivationViewSet(
     )
     @action(
         detail=False,
-        queryset=models.ActivationInstance.objects.order_by("id"),
+        queryset=models.RulebookProcess.objects.order_by("id"),
         filterset_class=filters.ActivationInstanceFilter,
         rbac_resource_type=ResourceType.ACTIVATION_INSTANCE,
         rbac_action=Action.READ,
@@ -155,8 +166,9 @@ class ActivationViewSet(
                 detail=f"Activation with ID={id} does not exist.",
             )
 
-        activation_instances = models.ActivationInstance.objects.filter(
-            activation_id=id
+        activation_instances = models.RulebookProcess.objects.filter(
+            activation_id=id,
+            parent_type=ProcessParentType.ACTIVATION,
         )
         filtered_instances = self.filter_queryset(activation_instances)
         result = self.paginate_queryset(filtered_instances)
@@ -224,7 +236,10 @@ class ActivationViewSet(
                 "modified_at",
             ]
         )
-        start_activation(activation_id=pk)
+        start_rulebook_process(
+            process_parent_type=ProcessParentType.ACTIVATION,
+            id=pk,
+        )
 
         return Response(status=status.HTTP_204_NO_CONTENT)
 
@@ -250,7 +265,10 @@ class ActivationViewSet(
             activation.save(
                 update_fields=["is_enabled", "status", "modified_at"]
             )
-            stop_activation(activation_id=activation.id)
+            stop_rulebook_process(
+                process_parent_type=ProcessParentType.ACTIVATION,
+                id=activation.id,
+            )
 
         return Response(status=status.HTTP_204_NO_CONTENT)
 
@@ -290,7 +308,10 @@ class ActivationViewSet(
                 {"errors": error}, status=status.HTTP_400_BAD_REQUEST
             )
 
-        restart_activation(activation_id=activation.id)
+        restart_rulebook_process(
+            process_parent_type=ProcessParentType.ACTIVATION,
+            id=activation.id,
+        )
 
         return Response(status=status.HTTP_204_NO_CONTENT)
 
@@ -332,7 +353,7 @@ class ActivationInstanceViewSet(
     viewsets.ReadOnlyModelViewSet,
     mixins.DestroyModelMixin,
 ):
-    queryset = models.ActivationInstance.objects.all()
+    queryset = models.RulebookProcess.objects.all()
     serializer_class = serializers.ActivationInstanceSerializer
     filter_backends = (defaultfilters.DjangoFilterBackend,)
     filterset_class = filters.ActivationInstanceFilter
@@ -358,22 +379,20 @@ class ActivationInstanceViewSet(
     )
     @action(
         detail=False,
-        queryset=models.ActivationInstanceLog.objects.order_by("id"),
+        queryset=models.RulebookProcessLog.objects.order_by("id"),
         filterset_class=filters.ActivationInstanceLogFilter,
         rbac_action=Action.READ,
         url_path="(?P<id>[^/.]+)/logs",
     )
     def logs(self, request, id):
-        instance_exists = models.ActivationInstance.objects.filter(
-            pk=id
-        ).exists()
+        instance_exists = models.RulebookProcess.objects.filter(pk=id).exists()
         if not instance_exists:
             raise api_exc.NotFound(
                 code=status.HTTP_404_NOT_FOUND,
                 detail=f"Activation Instance with ID={id} does not exist.",
             )
 
-        activation_instance_logs = models.ActivationInstanceLog.objects.filter(
+        activation_instance_logs = models.RulebookProcessLog.objects.filter(
             activation_instance_id=id
         ).order_by("id")
         activation_instance_logs = self.filter_queryset(
