@@ -37,6 +37,7 @@ class Credential(models.Model):
         default=CredentialType.REGISTRY,
     )
     username = models.TextField(null=True)
-    secret = EncryptedTextField(null=True)
+    secret = EncryptedTextField(null=False)
+    vault_identifier = models.TextField(null=True)
     created_at = models.DateTimeField(auto_now_add=True, null=False)
     modified_at = models.DateTimeField(auto_now=True, null=False)
