@@ -122,11 +122,12 @@ class Engine(ContainerEngine):
 
             log_handler.write("Starting Container", True)
             command = request.cmdline.command_and_args()
-            log_handler.write(f"Container args {command}", True)
+            command_log = request.cmdline.command_and_args(sanitized=True)
+            log_handler.write(f"Container args {command_log}", True)
             pod_args = self._load_pod_args(request)
             LOGGER.info(
                 "Creating container: "
-                f"command: {command}, "
+                f"command: {command_log}, "
                 f"pod_args: {pod_args}"
             )
             container = self.client.containers.run(
@@ -145,7 +146,7 @@ class Engine(ContainerEngine):
                 f"id: {container.id}, "
                 f"ports: {container.ports}, "
                 f"status: {container.status}, "
-                f"command: {command}, "
+                f"command: {command_log}, "
                 f"pod_args: {pod_args}"
             )
 
@@ -247,24 +248,36 @@ class Engine(ContainerEngine):
             since = None
             log_read_at = log_handler.get_log_read_at()
             if log_read_at:
-                since = int(log_handler.get_log_read_at().timestamp()) + 1
+                since = int(log_handler.get_log_read_at().timestamp())
 
-            container = self.client.containers.get(container_id)
             log_args = {"timestamps": True, "stderr": True}
+            last_timestamp = None
+            num_wrote_lines = 0
             if since:
                 log_args["since"] = since
-            timestamp = None
-            for logline in container.logs(**log_args):
+                num_wrote_lines = log_handler.num_log_write_from(since)
+
+            container = self.client.containers.get(container_id)
+
+            for i, logline in enumerate(container.logs(**log_args)):
+                if i < num_wrote_lines:
+                    continue
+
                 log = logline.decode("utf-8").strip()
                 log_parts = log.split(" ", 1)
-                timestamp = log_parts[0]
+                last_timestamp = log_parts[0]
                 if len(log_parts) > 1:
                     log_handler.write(
-                        lines=log_parts[1], flush=False, timestamp=False
+                        lines=log_parts[1],
+                        flush=False,
+                        timestamp=False,
+                        log_timestamp=int(
+                            parser.parse(last_timestamp).timestamp()
+                        ),
                     )
 
-            if timestamp:
-                dt = parser.parse(timestamp)
+            if last_timestamp:
+                dt = parser.parse(last_timestamp)
                 log_handler.flush()
                 log_handler.set_log_read_at(dt)
 
