@@ -75,14 +75,14 @@ class InitData:
 @pytest.mark.django_db
 def test_list_activation_instances(client: APIClient):
     activation = prepare_init_data()
-    instances = models.ActivationInstance.objects.bulk_create(
+    instances = models.RulebookProcess.objects.bulk_create(
         [
-            models.ActivationInstance(
+            models.RulebookProcess(
                 name="test-activation-instance-1",
                 activation=activation,
                 git_hash=PROJECT_GIT_HASH,
             ),
-            models.ActivationInstance(
+            models.RulebookProcess(
                 name="test-activation-instance-1",
                 activation=activation,
                 git_hash=PROJECT_GIT_HASH,
@@ -97,14 +97,14 @@ def test_list_activation_instances(client: APIClient):
 @pytest.mark.django_db
 def test_list_activation_instances_filter_name(client: APIClient):
     activation = prepare_init_data()
-    instances = models.ActivationInstance.objects.bulk_create(
+    instances = models.RulebookProcess.objects.bulk_create(
         [
-            models.ActivationInstance(
+            models.RulebookProcess(
                 name="activation-instance-1",
                 activation=activation,
                 git_hash=PROJECT_GIT_HASH,
             ),
-            models.ActivationInstance(
+            models.RulebookProcess(
                 name="test-activation-instance-2",
                 activation=activation,
                 git_hash=PROJECT_GIT_HASH,
@@ -124,15 +124,15 @@ def test_list_activation_instances_filter_name(client: APIClient):
 @pytest.mark.django_db
 def test_list_activation_instances_filter_status(client: APIClient):
     activation = prepare_init_data()
-    instances = models.ActivationInstance.objects.bulk_create(
+    instances = models.RulebookProcess.objects.bulk_create(
         [
-            models.ActivationInstance(
+            models.RulebookProcess(
                 name="activation-instance-1",
                 status="completed",
                 activation=activation,
                 git_hash=PROJECT_GIT_HASH,
             ),
-            models.ActivationInstance(
+            models.RulebookProcess(
                 name="test-activation-instance-2",
                 status="failed",
                 activation=activation,
@@ -153,7 +153,7 @@ def test_list_activation_instances_filter_status(client: APIClient):
 @pytest.mark.django_db
 def test_retrieve_activation_instance(client: APIClient):
     activation = prepare_init_data()
-    instance = models.ActivationInstance.objects.create(
+    instance = models.RulebookProcess.objects.create(
         name="activation-instance",
         activation=activation,
     )
@@ -172,7 +172,7 @@ def test_retrieve_activation_instance_not_exist(client: APIClient):
 @pytest.mark.django_db
 def test_delete_activation_instance(client: APIClient):
     activation = prepare_init_data()
-    instance = models.ActivationInstance.objects.create(
+    instance = models.RulebookProcess.objects.create(
         name="activation-instance",
         activation=activation,
     )
@@ -182,27 +182,25 @@ def test_delete_activation_instance(client: APIClient):
     )
     assert response.status_code == status.HTTP_204_NO_CONTENT
 
-    assert (
-        models.ActivationInstance.objects.filter(pk=instance.id).count() == 0
-    )
+    assert models.RulebookProcess.objects.filter(pk=instance.id).count() == 0
 
 
 @pytest.mark.django_db
 def test_list_logs_from_activation_instance(client: APIClient):
     activation = prepare_init_data()
-    instance = models.ActivationInstance.objects.create(
+    instance = models.RulebookProcess.objects.create(
         name="test-activation-instance",
         activation=activation,
     )
 
-    models.ActivationInstanceLog.objects.bulk_create(
+    models.RulebookProcessLog.objects.bulk_create(
         [
-            models.ActivationInstanceLog(
+            models.RulebookProcessLog(
                 log="activation-instance-log-1",
                 line_number=1,
                 activation_instance=instance,
             ),
-            models.ActivationInstanceLog(
+            models.RulebookProcessLog(
                 log="activation-instance-log-2",
                 line_number=2,
                 activation_instance=instance,
@@ -222,6 +220,7 @@ def test_list_logs_from_activation_instance(client: APIClient):
         "id",
         "line_number",
         "log",
+        "log_timestamp",
         "activation_instance",
     ]
 
@@ -229,19 +228,19 @@ def test_list_logs_from_activation_instance(client: APIClient):
 @pytest.mark.django_db
 def test_list_activation_instance_logs_filter(client: APIClient):
     activation = prepare_init_data()
-    instance = models.ActivationInstance.objects.create(
+    instance = models.RulebookProcess.objects.create(
         name="test-activation-instance",
         activation=activation,
     )
 
-    instance_logs = models.ActivationInstanceLog.objects.bulk_create(
+    instance_logs = models.RulebookProcessLog.objects.bulk_create(
         [
-            models.ActivationInstanceLog(
+            models.RulebookProcessLog(
                 log="activation-instance-log-1",
                 line_number=1,
                 activation_instance=instance,
             ),
-            models.ActivationInstanceLog(
+            models.RulebookProcessLog(
                 log="activation-instance-log-2",
                 line_number=2,
                 activation_instance=instance,
@@ -262,7 +261,7 @@ def test_list_activation_instance_logs_filter(client: APIClient):
 @pytest.mark.django_db
 def test_list_activation_instance_logs_filter_non_existent(client: APIClient):
     activation = prepare_init_data()
-    instance = models.ActivationInstance.objects.create(
+    instance = models.RulebookProcess.objects.create(
         name="test-activation-instance",
         activation=activation,
     )
@@ -278,7 +277,7 @@ def test_list_activation_instance_logs_filter_non_existent(client: APIClient):
 
 
 def assert_activation_instance_data(
-    data: Dict[str, Any], instance: models.ActivationInstance
+    data: Dict[str, Any], instance: models.RulebookProcess
 ):
     assert data == {
         "id": instance.id,
@@ -289,6 +288,7 @@ def assert_activation_instance_data(
         "started_at": instance.started_at.strftime(DATETIME_FORMAT),
         "ended_at": instance.ended_at,
         "status_message": ACTIVATION_STATUS_MESSAGE_MAP[instance.status],
+        "event_stream_id": None,
     }
 
 
