@@ -50,7 +50,7 @@ from aap_eda.services.activation.engine.kubernetes import (
 @dataclass
 class InitData:
     activation: models.Activation
-    activation_instance: models.ActivationInstance
+    activation_instance: models.RulebookProcess
 
 
 @pytest.fixture()
@@ -66,7 +66,7 @@ def init_data():
         name="activation",
         user=user,
     )
-    activation_instance = models.ActivationInstance.objects.create(
+    activation_instance = models.RulebookProcess.objects.create(
         name="test-instance",
         log_read_at=parser.parse("2023-10-30T19:18:48.362883381Z"),
         activation=activation,
@@ -82,6 +82,9 @@ def get_ansible_rulebook_cmdline(data: InitData):
     return AnsibleRulebookCmdLine(
         ws_url="ws://localhost:8000/api/eda/ws/ansible-rulebook",
         ws_ssl_verify="no",
+        ws_token_url="http://localhost:8000/api/eda/v1/auth/token/refresh",
+        ws_access_token="access",
+        ws_refresh_token="refresh",
         id=data.activation.id,
         log_level="-v",
         heartbeat=5,
@@ -98,7 +101,6 @@ def get_request(data: InitData):
         credential=Credential(username="admin", secret="secret"),
         ports=[("localhost", 8080)],
         mem_limit="8G",
-        mounts={"/dev": "/opt"},
         env_vars={"a": 1},
         extra_args={"b": 2},
     )
@@ -340,8 +342,8 @@ def test_engine_start(init_data, kubernetes_engine):
         f"{init_data.activation_instance.id}"
     )
 
-    assert models.ActivationInstanceLog.objects.count() == 5
-    assert models.ActivationInstanceLog.objects.last().log.endswith(
+    assert models.RulebookProcessLog.objects.count() == 5
+    assert models.RulebookProcessLog.objects.last().log.endswith(
         f"Job {engine.job_name} is running"
     )
 
@@ -378,7 +380,7 @@ def test_engine_start_with_pod_status(init_data, kubernetes_engine):
             watcher.stream.return_value = get_stream_event(phase)
             engine.start(request, log_handler)
 
-            assert models.ActivationInstanceLog.objects.last().log.endswith(
+            assert models.RulebookProcessLog.objects.last().log.endswith(
                 f"Job {engine.job_name} is running"
             )
 
@@ -479,7 +481,7 @@ def test_delete_secret(init_data, kubernetes_engine):
         core_api_mock.delete_namespaced_secret.return_value = status_mock
         engine._delete_secret(log_handler)
 
-        assert models.ActivationInstanceLog.objects.last().log.endswith(
+        assert models.RulebookProcessLog.objects.last().log.endswith(
             f"Secret {engine.secret_name} is deleted."
         )
 
@@ -487,7 +489,7 @@ def test_delete_secret(init_data, kubernetes_engine):
         status_mock.reason = "Not found"
         engine._delete_secret(log_handler)
 
-        assert models.ActivationInstanceLog.objects.last().log.endswith(
+        assert models.RulebookProcessLog.objects.last().log.endswith(
             f"Failed to delete secret {engine.secret_name}: "
             f"status: {status_mock.status}"
             f"reason: {status_mock.reason}"
@@ -520,7 +522,7 @@ def test_delete_service(init_data, kubernetes_engine):
 
         core_api_mock.delete_namespaced_service.assert_called_once()
 
-        assert models.ActivationInstanceLog.objects.last().log.endswith(
+        assert models.RulebookProcessLog.objects.last().log.endswith(
             f"Service {service_name} is deleted."
         )
 
@@ -558,7 +560,7 @@ def test_delete_job(init_data, kubernetes_engine):
         batch_api_mock.list_namespaced_job.return_value.items = None
         engine._delete_job(log_handler)
 
-        assert models.ActivationInstanceLog.objects.last().log.endswith(
+        assert models.RulebookProcessLog.objects.last().log.endswith(
             f"Job for {job_name} has been removed."
         )
 
@@ -604,7 +606,7 @@ def test_cleanup(init_data, kubernetes_engine):
                 services_mock.assert_called_once()
                 job_mock.assert_called_once()
 
-    assert models.ActivationInstanceLog.objects.last().log.endswith(
+    assert models.RulebookProcessLog.objects.last().log.endswith(
         f"Job {job_name} is cleaned up."
     )
 
@@ -642,9 +644,7 @@ def test_update_logs(init_data, kubernetes_engine):
             ]
             engine.update_logs(job_name, log_handler)
 
-            assert (
-                models.ActivationInstanceLog.objects.last().log == f"{message}"
-            )
+            assert models.RulebookProcessLog.objects.last().log == f"{message}"
             init_data.activation_instance.refresh_from_db()
             assert init_data.activation_instance.log_read_at > init_log_read_at
 
@@ -665,7 +665,7 @@ def test_update_logs(init_data, kubernetes_engine):
         with mock.patch.object(engine.client, "core_api") as core_api_mock:
             engine.update_logs(job_name, log_handler)
             msg = f"Pod with label {job_name} has unhandled state:"
-            assert msg in models.ActivationInstanceLog.objects.last().log
+            assert msg in models.RulebookProcessLog.objects.last().log
 
 
 @pytest.mark.django_db
