@@ -48,7 +48,7 @@ DATA_DIR = Path(__file__).parent / "data"
 @dataclass
 class InitData:
     activation: models.Activation
-    activation_instance: models.ActivationInstance
+    activation_instance: models.RulebookProcess
 
 
 @pytest.fixture()
@@ -64,7 +64,7 @@ def init_data():
         name="activation",
         user=user,
     )
-    activation_instance = models.ActivationInstance.objects.create(
+    activation_instance = models.RulebookProcess.objects.create(
         name="test-instance",
         log_read_at=parser.parse("2023-10-30T19:18:48.362883381Z"),
         activation=activation,
@@ -80,6 +80,9 @@ def get_ansible_rulebook_cmdline(data: InitData):
     return AnsibleRulebookCmdLine(
         ws_url="ws://localhost:8000/api/eda/ws/ansible-rulebook",
         ws_ssl_verify="no",
+        ws_token_url="http://localhost:8000/api/eda/v1/auth/token/refresh",
+        ws_access_token="access",
+        ws_refresh_token="refresh",
         id=data.activation.id,
         log_level="-v",
         heartbeat=5,
@@ -95,7 +98,7 @@ def get_request(data: InitData):
         cmdline=get_ansible_rulebook_cmdline(data),
         ports=[("localhost", 8080)],
         mem_limit="8G",
-        mounts={"/dev": "/opt"},
+        mounts=[{"/dev": "/opt"}],
         env_vars={"a": 1},
         extra_args={"b": 2},
     )
@@ -151,9 +154,15 @@ def test_get_podman_client(settings):
         assert client.api.base_url.netloc == "%2Frun%2Fpodman%2Fpodman.sock"
 
     client = get_podman_client()
+    xdg_runtime_dir = os.getenv(
+        "XDG_RUNTIME_DIR", f"%2Frun%2Fuser%2F{os.getuid()}"
+    )
+    # Replace any '/'s from XDG_RUNTIME_DIR.
+    xdg_runtime_dir = xdg_runtime_dir.replace("/", "%2F")
+
     assert (
         client.api.base_url.netloc
-        == f"%2Frun%2Fuser%2F{os.getuid()}%2Fpodman%2Fpodman.sock"
+        == f"{xdg_runtime_dir}%2Fpodman%2Fpodman.sock"
     )
 
 
@@ -196,10 +205,8 @@ def test_engine_start(init_data, podman_engine):
     engine.start(request, log_handler)
 
     engine.client.containers.run.assert_called_once()
-    assert models.ActivationInstanceLog.objects.count() == 4
-    assert models.ActivationInstanceLog.objects.last().log.endswith(
-        "is started."
-    )
+    assert models.RulebookProcessLog.objects.count() == 4
+    assert models.RulebookProcessLog.objects.last().log.endswith("is started.")
 
 
 @pytest.mark.django_db
@@ -243,7 +250,7 @@ def test_engine_start_with_credential(init_data, podman_engine):
         name=request.name,
         ports={"8080/tcp": 8080},
         mem_limit="8G",
-        mounts={"/dev": "/opt"},
+        mounts=request.mounts,
         environment={"a": 1},
         b=2,
     )
@@ -288,7 +295,7 @@ def test_engine_start_with_image_not_found_exception(init_data, podman_engine):
     ):
         engine.start(request, log_handler)
 
-    assert models.ActivationInstanceLog.objects.last().log.endswith(
+    assert models.RulebookProcessLog.objects.last().log.endswith(
         f"Image {request.image_url} not found"
     )
 
@@ -333,7 +340,7 @@ def test_engine_start_with_image_pull_exception(init_data, podman_engine):
         with pytest.raises(ContainerImagePullError, match=msg):
             engine.start(request, log_handler)
 
-    assert models.ActivationInstanceLog.objects.last().log.endswith(msg)
+    assert models.RulebookProcessLog.objects.last().log.endswith(msg)
 
 
 @pytest.mark.django_db
@@ -361,7 +368,7 @@ def test_engine_start_with_containers_run_exception(init_data, podman_engine):
 
     assert (
         "Container Start Error:"
-        in models.ActivationInstanceLog.objects.last().log
+        in models.RulebookProcessLog.objects.last().log
     )
 
 
@@ -453,7 +460,7 @@ def test_engine_cleanup(init_data, podman_engine):
 
     engine.cleanup("100", log_handler)
 
-    assert models.ActivationInstanceLog.objects.last().log.endswith(
+    assert models.RulebookProcessLog.objects.last().log.endswith(
         "Container 100 is cleaned up."
     )
 
@@ -472,7 +479,7 @@ def test_engine_cleanup_with_not_found_exception(init_data, podman_engine):
 
     engine.cleanup("100", log_handler)
 
-    assert models.ActivationInstanceLog.objects.last().log.endswith(
+    assert models.RulebookProcessLog.objects.last().log.endswith(
         "Container 100 not found."
     )
 
@@ -546,10 +553,10 @@ def test_engine_update_logs(init_data, podman_engine):
 
     engine.update_logs("100", log_handler)
 
-    assert models.ActivationInstanceLog.objects.count() == len(
+    assert models.RulebookProcessLog.objects.count() == len(
         container_mock.logs.return_value
     )
-    assert models.ActivationInstanceLog.objects.last().log == f"{message}"
+    assert models.RulebookProcessLog.objects.last().log == f"{message}"
 
     init_data.activation_instance.refresh_from_db()
     assert init_data.activation_instance.log_read_at > init_log_read_at
@@ -563,7 +570,7 @@ def test_engine_update_logs_with_container_not_found(init_data, podman_engine):
     engine.client.containers.exists.return_value = None
     engine.update_logs("100", log_handler)
 
-    assert models.ActivationInstanceLog.objects.last().log.endswith(
+    assert models.RulebookProcessLog.objects.last().log.endswith(
         "Container 100 not found."
     )
 
@@ -589,9 +596,11 @@ def test_set_auth_json(podman_engine):
     with mock.patch("os.path.dirname"):
         engine._set_auth_json_file()
 
-        assert (
-            engine.auth_file == f"/run/user/{os.getuid()}/containers/auth.json"
+        xdg_runtime_dir = os.getenv(
+            "XDG_RUNTIME_DIR", f"/run/user/{os.getuid()}"
         )
+
+        assert engine.auth_file == f"{xdg_runtime_dir}/containers/auth.json"
 
 
 @pytest.mark.django_db
