@@ -17,7 +17,7 @@ from unittest import mock
 import pytest
 
 from aap_eda.core import models
-from aap_eda.core.enums import ActivationRequest
+from aap_eda.core.enums import ActivationRequest, ProcessParentType
 from aap_eda.services.activation.restart_helper import (
     _queue_auto_start,
     system_restart_activation,
@@ -42,11 +42,12 @@ def activation():
 @pytest.mark.django_db
 @mock.patch("aap_eda.services.activation.restart_helper.enqueue_delay")
 def test_system_restart_activation(enqueue_mock, activation):
-    system_restart_activation(activation.id, 5)
+    system_restart_activation(ProcessParentType.ACTIVATION, activation.id, 5)
     enqueue_args = [
         "default",
         5,
         _queue_auto_start,
+        ProcessParentType.ACTIVATION,
         activation.id,
     ]
     enqueue_mock.assert_called_once_with(*enqueue_args)
@@ -54,8 +55,9 @@ def test_system_restart_activation(enqueue_mock, activation):
 
 @pytest.mark.django_db
 def test_queue_auto_start(activation):
-    _queue_auto_start(activation.id)
+    _queue_auto_start(ProcessParentType.ACTIVATION, activation.id)
 
     queued = models.ActivationRequestQueue.objects.first()
-    assert queued.activation == activation
+    assert queued.process_parent_type == ProcessParentType.ACTIVATION
+    assert queued.process_parent_id == activation.id
     assert queued.request == ActivationRequest.AUTO_START
